[OPP-1497] fjerne gamle soap-integrasjonen
Mulig git-diffen blir litt rar her ettersom at den gamle soap-integrasjonen er fjernet, og alt som het "rest"-ettellerannet har blitt renamet.
