+++ .nais/prod.yml
@@ -86,8 +86,6 @@ spec:
       value: "https://digdir-krr-proxy.intern.nav.no"
     - name: DIGDIR_KRR_SCOPE
       value: "prod-gcp:team-rocket:digdir-krr-proxy"
-    - name: UTBETALING_V1_URL
-      value: "https://wasapp.adeo.no/nav-utbetaldata-ws/virksomhet/Utbetaling_v1"
     - name: UTBETALDATA_SOKOS_URL
       value: "https://sokos-utbetaldata.intern.nav.no"
     - name: UTBETAL_SOKOS_SCOPE
