+++ src/main/kotlin/no/nav/Env.kt
@@ -19,7 +19,6 @@ interface Env {
     val pdlScope: DownstreamApi
     val digdirUrl: String
     val digdirScope: DownstreamApi
-    val utbetalingerUrl: String
     val utbetalingSokosUrl: String
     val utbetalingSokosScope: DownstreamApi
     val safUrl: String
@@ -42,7 +41,6 @@ class EnvImpl : Env {
     override val pdlScope: DownstreamApi = getRequiredConfig("PDL_SCOPE").toDownstreamApi()
     override val digdirUrl: String = getRequiredConfig("DIGDIR_KRR_URL")
     override val digdirScope: DownstreamApi = getRequiredConfig("DIGDIR_KRR_SCOPE").toDownstreamApi()
-    override val utbetalingerUrl: String = getRequiredConfig("UTBETALING_V1_URL")
     override val utbetalingSokosUrl: String = getRequiredConfig("UTBETALDATA_SOKOS_URL")
     override val utbetalingSokosScope: DownstreamApi = getRequiredConfig("UTBETAL_SOKOS_SCOPE").toDownstreamApi()
     override val safUrl: String = getRequiredConfig("SAF_URL")
